+++ src/lxp/mod.rs
@@ -1,3 +1,4 @@
 pub mod inverter;
 pub mod packet;
 pub mod packet_decoder;
+pub mod register_parser;