+++ CHANGELOG.md
@@ -1,11 +1,20 @@
 # Unreleased
 
+This release deprecates the MQTT topics `lxp/*/inputs/1` (and 2, 3, 4, all). Instead, new applications should switch to using `lxp/*/input/v_bat/parsed` (where `v_bat` is the JSON Key from the old messages). This solves the problem of which messages to publish, and when (inputs/all is particularly problematic!). For more, see https://github.com/celsworth/lxp-bridge/discussions/262
+
+
 * Reconnect to inverter after 15 minutes of not receiving any data (#223)
 * Fix max/min cell temperature/voltage decoding as reported from BMS (#227)
 * Add more HA entities: max/min cell temp/voltage, more charge powers (#228)
 * Add ReadInput4 with EG4 18k generator data (#239, @pmccut)
 * Add ReadInput4 keys to HA discovery (#240, @jgulick48)
 * Fix min_chg_curr/max_chg_curr decoding in ReadInputAll packet (#242, @presto8)
+* Cache Hold/Input registers internally as they're seen for later use (#248)
+* Remove publish_individual_input configuration option (now they always are) (#250)
+* Convert HA to use individual inputs messages (#253)
+* New configuration open publish_inputs_all_trigger (#256)
+* Move time register messages into new parser (start+end times for AC Charge etc) (#259)
+* Publish start-end time message when time registers change externally (#261)
 
 
 # 0.13.0 - 27th October 2023